requirements.txt
PyQt6>=6.9.0
Pillow>=11.0.0
pyinstaller>=6.0.0
